move operator out of ConditionVerifier class
